new themes and theme obj change
