+++ components/screens/post/NewPostScreen.tsx
@@ -121,7 +121,7 @@ function NewPostScreen({
             style={[
               styles.input,
               {
-                backgroundColor: theme.colors.app.bgTertiary,
+                backgroundColor: theme.colors.app.bg,
                 color: theme.colors.app.textPrimary,
               },
             ]}
