+++ components/ui/buttons/ImageButton.tsx
@@ -41,7 +41,7 @@ function ImageButton({ src }: ImageButtonProps) {
     <>
       <Pressable onPress={onPress}>
         <HStack
-          backgroundColor={theme.colors.app.bgTertiary}
+          backgroundColor={theme.colors.app.bg}
           borderRadius={5}
           padding={2}
           flexDirection="row"
