+++ theme/darkColors.ts
@@ -1,0 +1,377 @@
+import { IColors } from "./common";
+
+export const darkThemeColors: IColors = {
+  accent: "#2C9AFF",
+  accentHighlight: "#1F2E45",
+
+  textPrimary: "#FFFFFF",
+  textSecondary: "#B8B9C1",
+
+  fg: "#1B1B1B",
+  bg: "#101010",
+  navBarBg: "#1B1B1B",
+
+  border: "#242424",
+
+  upvote: "#EE923D",
+  upvoteText: "#FFF",
+  downvote: "#1A84E5",
+  downvoteText: "#FFF",
+
+  success: "#48C10F",
+  successBg: "#D7F8DA",
+  successBorder: "#2B8302",
+  error: "#F8372B",
+  errorBg: "#FFC6C2",
+  errorBorder: "#A62525",
+  warn: "#DEA435",
+  warnBg: "#FFEDD9",
+  warnBorder: "#CB6A11",
+  info: "#3398f2",
+  infoBg: "#CAE6FF",
+  infoBorder: "#316677",
+
+  users: {
+    mod: "#54CB60",
+    admin: "#FF4848",
+    dev: "#CB66FA",
+    op: "#54CB60",
+    me: "#FFB800",
+  },
+
+  comments: {
+    1: "#FA6969",
+    2: "#FFB800",
+    3: "#FFE500",
+    4: "#36CF3C",
+    5: "#4BA9FF",
+  },
+
+  inputBg: "#2F2F2F",
+
+  bookmark: "#48C10F",
+  bookmarkText: "#fff",
+};
+
+export const moodyPurpleThemeColors: IColors = {
+  accent: "#8877F1",
+  accentHighlight: "#352945",
+
+  textPrimary: "#EDE1FF",
+  textSecondary: "#AEB5ED",
+  // textTertiary: "#AEB5ED",
+
+  fg: "#1E1529",
+  bg: "#120E1D",
+  navBarBg: "#1D1528",
+  // bgSecondary: "#1E1529",
+  // bgTertiary: "#392B4A",
+
+  border: "#3C2B3D",
+
+  upvote: "#EE923D",
+  upvoteText: "#FFF",
+  downvote: "#1A84E5",
+  downvoteText: "#FFF",
+
+  success: "#48C10F",
+  successBg: "#D7F8DA",
+  successBorder: "#2B8302",
+  error: "#F8372B",
+  errorBg: "#FFC6C2",
+  errorBorder: "#A62525",
+  warn: "#DEA435",
+  warnBg: "#FFEDD9",
+  warnBorder: "#CB6A11",
+  info: "#3398f2",
+  infoBg: "#CAE6FF",
+  infoBorder: "#316677",
+
+  users: {
+    mod: "#54CB60",
+    admin: "#FF4848",
+    dev: "#CB66FA",
+    op: "#54CB60",
+    me: "#FFB800",
+  },
+
+  comments: {
+    1: "#FA6969",
+    2: "#FFB800",
+    3: "#FFE500",
+    4: "#36CF3C",
+    5: "#4BA9FF",
+  },
+
+  inputBg: "#2D2736",
+
+  bookmark: "#48C10F",
+  bookmarkText: "#fff",
+};
+
+export const draculaThemeColors: IColors = {
+  accent: "#ff79c6",
+  accentHighlight: "#bd93f9",
+
+  textPrimary: "#F8F8F2",
+  textSecondary: "#b8b8bf",
+
+  fg: "#232635",
+  bg: "#1B1D29",
+  navBarBg: "#10141C",
+  // bgSecondary: "#282a36",
+  // bgTertiary: "#10141C",
+
+  border: "#44475a",
+
+  upvote: "#FF9580",
+  upvoteText: "#FFF",
+  downvote: "#9580FF",
+  downvoteText: "#FFF",
+
+  success: "#48C10F",
+  successBg: "#D7F8DA",
+  successBorder: "#2B8302",
+  error: "#F8372B",
+  errorBg: "#FFC6C2",
+  errorBorder: "#A62525",
+  warn: "#DEA435",
+  warnBg: "#FFEDD9",
+  warnBorder: "#CB6A11",
+  info: "#3398f2",
+  infoBg: "#CAE6FF",
+  infoBorder: "#316677",
+
+  users: {
+    mod: "#50fa7b",
+    admin: "#ff5555",
+    dev: "#bd93f9",
+    op: "#8be9fd",
+    me: "#ffb86c",
+  },
+
+  comments: {
+    1: "#ff79c6",
+    2: "#ffb86c",
+    3: "#f1fa8c",
+    4: "#50fa7b",
+    5: "#8be9fd",
+  },
+
+  inputBg: "#252837",
+
+  bookmark: "#48C10F",
+  bookmarkText: "#fff",
+};
+
+export const nightOwlThemeColors: IColors = {
+  accent: "#544AC6",
+  accentHighlight: "#1F1932",
+
+  textPrimary: "#C2B8DF",
+  textSecondary: "#8B7C92",
+
+  fg: "#090909",
+  bg: "#000000",
+  navBarBg: "#090909",
+
+  border: "#241543",
+
+  upvote: "#EE923D",
+  upvoteText: "#FFF",
+  downvote: "#1A84E5",
+  downvoteText: "#FFF",
+
+  success: "#48C10F",
+  successBg: "#D7F8DA",
+  successBorder: "#2B8302",
+  error: "#F8372B",
+  errorBg: "#FFC6C2",
+  errorBorder: "#A62525",
+  warn: "#DEA435",
+  warnBg: "#FFEDD9",
+  warnBorder: "#CB6A11",
+  info: "",
+  infoBg: "#CAE6FF",
+  infoBorder: "#316677",
+
+  users: {
+    mod: "#54CB60",
+    admin: "#FF4848",
+    dev: "#CB66FA",
+    op: "#54CB60",
+    me: "#FFB800",
+  },
+
+  comments: {
+    1: "#FA6969",
+    2: "#FFB800",
+    3: "#FFE500",
+    4: "#36CF3C",
+    5: "#4BA9FF",
+  },
+
+  inputBg: "#221C2F",
+
+  bookmark: "#48C10F",
+  bookmarkText: "#fff",
+};
+
+export const sleepyRainforestThemeColors: IColors = {
+  accent: "#70B944",
+  accentHighlight: "#2A382B",
+
+  textPrimary: "#E3FFE2",
+  textSecondary: "#DBFFCE",
+
+  fg: "#161D14",
+  bg: "#0A1308",
+  navBarBg: "#171D15",
+
+  border: "#142811",
+
+  upvote: "#EE923D",
+  upvoteText: "#FFF",
+  downvote: "#1A84E5",
+  downvoteText: "#FFF",
+
+  success: "#48C10F",
+  successBg: "#D7F8DA",
+  successBorder: "#2B8302",
+  error: "#F8372B",
+  errorBg: "#FFC6C2",
+  errorBorder: "#A62525",
+  warn: "#DEA435",
+  warnBg: "#FFEDD9",
+  warnBorder: "#CB6A11",
+  info: "",
+  infoBg: "#CAE6FF",
+  infoBorder: "#316677",
+
+  users: {
+    mod: "#54CB60",
+    admin: "#FF4848",
+    dev: "#CB66FA",
+    op: "#54CB60",
+    me: "#FFB800",
+  },
+
+  comments: {
+    1: "#FA6969",
+    2: "#FFB800",
+    3: "#FFE500",
+    4: "#36CF3C",
+    5: "#4BA9FF",
+  },
+
+  inputBg: "#222B1F",
+
+  bookmark: "#48C10F",
+  bookmarkText: "#fff",
+};
+
+export const goldenHourThemeColors: IColors = {
+  accent: "#C7804C",
+  accentHighlight: "#382917",
+
+  textPrimary: "#FFEEEE",
+  textSecondary: "#FFD1A7",
+
+  fg: "#1D1106",
+  bg: "#110404",
+  navBarBg: "#1B1207",
+
+  border: "#402512",
+
+  upvote: "#EE923D",
+  upvoteText: "#FFF",
+  downvote: "#1A84E5",
+  downvoteText: "#FFF",
+
+  success: "#48C10F",
+  successBg: "#D7F8DA",
+  successBorder: "#2B8302",
+  error: "#F8372B",
+  errorBg: "#FFC6C2",
+  errorBorder: "#A62525",
+  warn: "#DEA435",
+  warnBg: "#FFEDD9",
+  warnBorder: "#CB6A11",
+  info: "",
+  infoBg: "#CAE6FF",
+  infoBorder: "#316677",
+
+  users: {
+    mod: "#54CB60",
+    admin: "#FF4848",
+    dev: "#CB66FA",
+    op: "#54CB60",
+    me: "#FFB800",
+  },
+
+  comments: {
+    1: "#FA6969",
+    2: "#FFB800",
+    3: "#FFE500",
+    4: "#36CF3C",
+    5: "#4BA9FF",
+  },
+
+  inputBg: "#302211",
+
+  bookmark: "#48C10F",
+  bookmarkText: "#fff",
+};
+
+export const nauticalThemeColors: IColors = {
+  accent: "#2C9AFF",
+  accentHighlight: "#1F2E45",
+
+  textPrimary: "#E2EEFF",
+  textSecondary: "#90B6FF",
+
+  fg: "#111626",
+  bg: "#0C111A",
+  navBarBg: "#121625",
+
+  border: "#111626",
+
+  upvote: "#EE923D",
+  upvoteText: "#FFF",
+  downvote: "#1A84E5",
+  downvoteText: "#FFF",
+
+  success: "#48C10F",
+  successBg: "#D7F8DA",
+  successBorder: "#2B8302",
+  error: "#F8372B",
+  errorBg: "#FFC6C2",
+  errorBorder: "#A62525",
+  warn: "#DEA435",
+  warnBg: "#FFEDD9",
+  warnBorder: "#CB6A11",
+  info: "",
+  infoBg: "#CAE6FF",
+  infoBorder: "#316677",
+
+  users: {
+    mod: "#54CB60",
+    admin: "#FF4848",
+    dev: "#CB66FA",
+    op: "#54CB60",
+    me: "#FFB800",
+  },
+
+  comments: {
+    1: "#FA6969",
+    2: "#FFB800",
+    3: "#FFE500",
+    4: "#36CF3C",
+    5: "#4BA9FF",
+  },
+
+  inputBg: "#2A314A",
+
+  bookmark: "#48C10F",
+  bookmarkText: "#fff",
+};
