+++ theme/lightColors.ts
@@ -1,0 +1,55 @@
+import { IColors } from "./common";
+
+// eslint-disable-next-line import/prefer-default-export
+export const lightThemeColors: IColors = {
+  accent: "#2092FB",
+  accentHighlight: "#E7F2FF",
+
+  textPrimary: "#393939",
+  textSecondary: "#979BA2",
+
+  fg: "#FFFFFF",
+  bg: "#F2F2F2",
+  navBarBg: "#FEFEFE",
+
+  border: "#C8C8CA",
+
+  upvote: "#EE923D",
+  upvoteText: "#FFF",
+  downvote: "#1A84E5",
+  downvoteText: "#FFF",
+
+  success: "#48C10F",
+  successBg: "#D7F8DA",
+  successBorder: "#2B8302",
+  error: "#F8372B",
+  errorBg: "#FFC6C2",
+  errorBorder: "#A62525",
+  warn: "#DEA435",
+  warnBg: "#FFEDD9",
+  warnBorder: "#CB6A11",
+  info: "#3398f2",
+  infoBg: "#CAE6FF",
+  infoBorder: "#316677",
+
+  users: {
+    mod: "#54CB60",
+    admin: "#FF4848",
+    dev: "#CB66FA",
+    op: "#54CB60",
+    me: "#FFB800",
+  },
+
+  comments: {
+    1: "#FA6969",
+    2: "#FFB800",
+    3: "#FFE500",
+    4: "#36CF3C",
+    5: "#4BA9FF",
+  },
+
+  inputBg: "#FAFAFA",
+
+  bookmark: "#48C10F",
+  bookmarkText: "#fff",
+};
