+++ theme/themeOptions.ts
@@ -1,21 +1,25 @@
 import {
-  brownTheme,
+  moodyPurpleTheme,
   darkTheme,
   draculaTheme,
   draculaThemePurple,
   lightTheme,
   nightOwlTheme,
-  purpleTheme,
+  sleepyRainforestTheme,
+  goldenHourTheme,
+  nauticalTheme,
 } from "./theme";
 
 enum EThemeOptions {
-  BROWN = "Brown",
   LIGHT = "Light",
   DARK = "Dark",
-  PURPLE = "Purple",
   DRACULA = "Dracula",
   DRACULA_PURPLE = "Dracula (Purple)",
+  MOODY_PURPLE = "Moody Purple",
   NIGHT_OWL = "Night Owl",
+  SLEEPY_RAINFOREST = "Sleepy Rainforest",
+  GOLDEN_HOUR = "Golden Hour",
+  NAUTICAL = "Nautical",
 }
 
 export type ThemeOptions = `${EThemeOptions}`;
@@ -24,11 +28,13 @@ export const ThemeOptionsArr = Object.values(EThemeOptions);
 
 // TODO: should get the type figured out for theme obj
 export const ThemeOptionsMap: Record<ThemeOptions, any> = {
-  Brown: brownTheme,
-  Dark: darkTheme,
   Light: lightTheme,
-  Purple: purpleTheme,
+  Dark: darkTheme,
   Dracula: draculaTheme,
   "Dracula (Purple)": draculaThemePurple,
+  "Moody Purple": moodyPurpleTheme,
   "Night Owl": nightOwlTheme,
+  "Sleepy Rainforest": sleepyRainforestTheme,
+  "Golden Hour": goldenHourTheme,
+  Nautical: nauticalTheme,
 };
